+++ core-ui/src/shared/ResourceForm/components/CollapsibleSection.js
