+++ core-ui/src/shared/ResourceForm/fields/Editor.scss
