+++ core-ui/src/shared/ResourceForm/inputs/index.js
@@ -2,4 +2,5 @@ export * from './Text';
 export * from './Number';
 export * from './Switch';
 export * from './Dropdown';
-export * from '.Checkboxes';
+export * from './Checkboxes';
+export * from './ComboboxInput';
